update: update header components
